+++ python/target_selection/cartons/mwm_snc.py
@@ -8,6 +8,7 @@
 
 import random
 
+import peewee
 from peewee import fn
 
 from sdssdb.peewee.sdss5db.catalogdb import (
@@ -243,7 +244,7 @@ class MWM_SNC_100pc_BOSS_single_Carton(MWM_SNC_100pc_BOSS_Carton):
     is the cadence and priority.
 
     Metadata:
-    Priority: 1801
+    Priority: 1806
     Cadence:  if G < 16, bright_1x1 ; if G > 16, dark_1x1
     Instrument: BOSS
     can_offset = True
@@ -256,7 +257,7 @@ class MWM_SNC_100pc_BOSS_single_Carton(MWM_SNC_100pc_BOSS_Carton):
     mapper = "MWM"
     instrument = "BOSS"
     cadence = None  # cadence is set in post_process()
-    priority = 1801
+    priority = 1806
     can_offset = True
 
     def build_query(self, version_id, query_region=None):
@@ -631,3 +632,126 @@ class MWM_SNC_Ext_filler_BOSS_Carton(MWM_SNC_Ext_BOSS_Base_Carton):
                     + str(current_catalogid)
                     + ";"
                 )
+
+
+class MWM_SNC_100pc_bright_BOSS_single_Carton(MWM_SNC_100pc_BOSS_Carton):
+    """mwm_snc_100pc_bright_boss_single
+    Shorthand name: mwm_snc_100pc_bright_boss_single
+
+    Existing carton code:
+    https://github.com/sdss/target_selection/blob/main/python/
+    target_selection/cartons/mwm_snc.py Simplified Description of selection
+    criteria :  Select everything within 100pc (i.e. parallax > 10mas) and
+    Gaia-G fainter than 10th mag. Add some constraints on astrometric excess
+    noise in crowded regions (Galactic plane, LMC, SMC, towards Galactic
+    center). For details, see here. NOTE this is a subcarton of
+    mwm_snc_100pc_boss (see above) where the only change is the cadence and
+    priority. So, this should be able to inherit `class
+    MWM_SNC_100pc_BOSS_Carton()`
+
+    Return columns: Same columns as before + astrometric_excess_noise,
+    phot_bp_rp_excess_factor, ruwe, phot_bp_mean_mag , phot_rp_mean_mag
+
+    Metadata:
+    Priority: 1807
+    Cadence:  if G < 18, bright_1x1 ; if G > 18, dark_1x1
+    Instrument: BOSS
+    can_offset = True
+
+    Lead contact: Ilija Medan
+    """
+
+    name = "mwm_snc_100pc_bright_boss_single"
+    category = "science"
+    program = "mwm_snc"
+    mapper = "MWM"
+    instrument = "BOSS"
+    cadence = None  # cadence is set in post_process()
+    priority = 1807
+    can_offset = True
+
+    def post_process(self, model):
+        cursor = self.database.execute_sql(
+            "select catalogid, phot_g_mean_mag from "
+            + " sandbox.temp_mwm_snc_100pc_bright_boss_single "
+            + " order by catalogid;"
+        )
+
+        output = cursor.fetchall()
+
+        for i in range(len(output)):
+            current_catalogid = output[i][0]
+            current_phot_g_mean_mag = output[i][1]
+
+            if current_phot_g_mean_mag < 18:
+                self.database.execute_sql(
+                    " update sandbox.temp_mwm_snc_100pc_bright_boss_single "
+                    + " set cadence = 'bright_1x1' "
+                    + " where catalogid = "
+                    + str(current_catalogid)
+                    + ";"
+                )
+            else:
+                self.database.execute_sql(
+                    " update sandbox.temp_mwm_snc_100pc_bright_boss_single "
+                    + " set cadence = 'dark_1x1' "
+                    + " where catalogid = "
+                    + str(current_catalogid)
+                    + ";"
+                )
+
+
+class MWM_SNC_100pc_low_lum__BOSS_Carton(MWM_SNC_100pc_BOSS_Carton):
+    """mwm_snc_100pc_low_lum_boss_single
+    Shorthand name: mwm_snc_100pc_low_lum_boss_single
+
+    Existing carton code:
+    https://github.com/sdss/target_selection/blob/main/python/
+    target_selection/cartons/mwm_snc.py Simplified Description of selection
+    criteria :  Select everything within 100pc (i.e. parallax > 10mas) and
+    Gaia-G fainter than 10th mag. Add some constraints on astrometric excess
+    noise in crowded regions (Galactic plane, LMC, SMC, towards Galactic
+    center). For details, see here. NOTE this is a subcarton of
+    mwm_snc_100pc_boss (see above). So, this should be able to inherit
+    `class MWM_SNC_100pc_BOSS_Carton()` and then add the below condition in
+    red.
+
+    Pseudo-code: The below text
+    is what will need to be added on top of the BOSS base carton
+    (`class MWM_SNC_100pc_BOSS_Carton()`)
+     & (phot_g_mean_mag + 5 * log10(0.001 * parallax) + 5 > 16)
+
+    Return columns: Same columns as before + astrometric_excess_noise,
+    phot_bp_rp_excess_factor, ruwe, phot_bp_mean_mag , phot_rp_mean_mag
+
+    Metadata:
+    Priority: 1310
+    Cadence:  dark_1x1
+    Instrument: BOSS
+    can_offset = True
+
+    Lead contact: Ilija Medan
+    """
+
+    name = "mwm_snc_100pc_low_lum_boss_single"
+    category = "science"
+    program = "mwm_snc"
+    mapper = "MWM"
+    instrument = "BOSS"
+    cadence = "dark_1x1"
+    priority = 1310
+    can_offset = True
+
+    def build_query(self, version_id, query_region=None):
+        query = super().build_query(version_id, query_region)
+        query = query.where(
+            Gaia_DR3.phot_g_mean_mag + 5 * peewee.fn.log10(0.001 * Gaia_DR3.parallax) + 5 > 16
+        )
+
+        return query
+
+    # We have set the cadence above so we do not want to run
+    # the post_process() of the base class MWM_SNC_100pc_BOSS_Carton.
+    # Hence, we have the below do-nothing post_process().
+    def post_process(self, model, **kwargs):
+        return model