+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 ## Next version
 
+# 1.3.14 - August 12 2024
+
+### 🚀 New
+
+* add cartons to target_selection plan 1.2.15
+
+## 1.3.13 - August 10, 2024
+
+### 🚀 New
+
+* add cartons to target_selection plan 1.2.14
+
+## 1.3.12 - August 1, 2024
+
+### 🚀 New
+
+* add cartons to target_selection plan 1.2.13
+
 ## 1.3.11 - August 1, 2024
 
 ### 🚀 New